pyproject.toml
[project]
name = "captionqa"
version = "0.1.0"
description = "Panoramic video captioning and interactive QA system"
authors = [
  { name = "Will Romano" },
  { name = "Ethan Baird" }
]
requires-python = ">=3.9"
dependencies = [
  # Core ML stack (Windows-friendly wheels)
  "torch>=2.2,<2.4",
  "torchvision>=0.17,<0.19",
  "torchaudio>=2.2,<2.4",
  # Transformers + tokenizers (avoid old Rust-build-only versions)
  "transformers>=4.40,<4.53",
  "tokenizers>=0.15,<0.20",
  # Vision/IO
  "opencv-python>=4.8,<4.10",
  "ffmpeg-python",
  # Audio + numerics
  "librosa",
  "numpy",
  # Utilities
  "pandas",
  "matplotlib",
  "scikit-learn",
  "tqdm",
  "requests",
  "einops",
  "sentencepiece",
  "accelerate",
  "peft",
  "huggingface-hub[hf_xet]>=0.23",
  "gdown",
  "datasets",
  # Metrics
  "sacrebleu>=2.3,<3.0",
  # Dev/test
  "pytest",
  "responses",
  "pytest-httpserver"
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.hatch.envs.default]
type = "uv"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest",
  "responses",
  "pytest-httpserver",
  "sacrebleu>=2.3,<3.0"
]
directml = [
  "torch-directml>=0.2.0; platform_system == 'Windows'"
]
